continued development oin cancellation bp
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C30423

+++ CDRAnalysis/Vanrise.Fzero.FraudAnalysis.Arguments/CancelStrategyExecutionProcessInput.cs
@@ -6,7 +6,7 @@ namespace Vanrise.Fzero.FraudAnalysis.BP.Arguments
 {
     public class CancelStrategyExecutionProcessInput : Vanrise.BusinessProcess.Entities.BaseProcessInputArgument
     {
-        public int StrategyExecutionId { get; set; }
+        public long StrategyExecutionId { get; set; }
         public override string GetTitle()
         {
             return String.Format("Cancel Strategy Execution Id: {0}", StrategyExecutionId);
